_examples/speech2text/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
nats-py==2.6.0
requests==2.31.0
SpeechRecognition==3.10.0
urllib3==2.1.0

